refactoring list and chat history nodes
